requirements.txt
requests
python-dotenv
pytest
requests
locust